refactor: move Vampire and Werewolf mixins outside of epicfight.mixins.json
