+++ src/main/resources/epicfight-compat.werewolves.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": false,
+  "minVersion": "0.8",
+  "compatibilityLevel": "JAVA_21",
+  "package": "yesman.epicfight.compat.werewolves.mixin",
+  "client": [
+    "MixinHumanWerewolfLayer"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
